requirements.txt
discord.py==2.3.2
aiohttp==3.8.5
aiosqlite==0.19.0
python-dotenv==1.0.0
